docs(date-picker): combine filter trigger button stories
